Add custom padding

+++ src/include/mml_pooling_node.hpp
@@ -82,8 +82,8 @@ public:
    */
   virtual void pooling(const shared_ptr<Tensor<T>> t,
                        array_mml<int> input_shape, array_mml<int> output_shape,
-                       vector<int> effective_kernel_shape, float pad_h,
-                       float pad_w, string auto_pad) = 0;
+                       vector<int> effective_kernel_shape, int pad_h, int pad_w,
+                       string auto_pad) = 0;
 
 protected:
   //--------Inputs----------
